feat: draw 2D circle
2D circle SDF is fairly easy to understand, displaying anything 3D
might be a no go in time. I think raymarching and 3D SDFs is the
way to go for 3D in the fragment shader, but still not 100% sure.

Leaving visuals at this for now, need to focus on web audio side
and making it possible to trigger the visuals plus audio based on
user interaction so that audio can play properly